[CL 13753156 by Marc Audy in ue5-main branch]

+++ Engine/Plugins/Runtime/AudioModulation/Source/AudioModulation/Public/SoundModulationValue.h
@@ -33,7 +33,7 @@ struct AUDIOMODULATION_API FSoundModulationValue
 	/** Current value lerping toward target */
 	float GetCurrentValue() const;
 
-	void Update(float Elapsed);
+	void Update(double Elapsed);
 
 private:
 	float Value;
